ml/requirements.txt
fastapi==0.111.0
numpy==2.0.0
opencv_python==4.9.0.80
Pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
uvicorn==0.30.1
